Extract SolArk client and add CLI
