+++ custom_components/solark/solark_errors.py
@@ -1,0 +1,5 @@
+"""SolArk shared exception types."""
+
+
+class SolArkCloudAPIError(Exception):
+    """Exception for Sol-Ark Cloud API errors."""
