+++ solark_cli/__init__.py
@@ -1,0 +1,1 @@
+"""SolArk CLI package."""
