+++ solark_cli/__main__.py
@@ -1,0 +1,6 @@
+"""Entrypoint for SolArk CLI."""
+
+from .cli import main
+
+if __name__ == "__main__":
+    main()
